+++ CHANGELOG.md
@@ -57,6 +57,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Exports `useMediaQuery` and `chakra` from Chakra UI.
 - Updates `Accordion` to close focused panel when "esc" key is pressed.
 - Updates the `TextInput` and `SearchBar` components to better associate the input element to the entire component's helper text.
+- Updates `Accordion` to close panel when element within panel is focused and "esc" key is pressed.
 
 ## 3.1.4 (May 23, 2024)